added validation and error message handling

+++ context/authContext.jsx
@@ -1,6 +1,7 @@
 import axios from "axios";
 import * as SecureStore from 'expo-secure-store';
 import { createContext, useContext, useEffect, useState } from "react";
+import { Alert } from "react-native";
 
 export const AuthContext = createContext()
 export const AuthContextProvider = ({ children }) => {
@@ -10,29 +11,33 @@ export const AuthContextProvider = ({ children }) => {
     const handleRegister = async (username, email, password) => {
 
         try {
-            const response = await axios.post(
+            const { data: userData } = await axios.post(
                 '/users/register',
                 {
                     username,
                     email,
                     password
+                },
+                {
+                    timeout: 5000
                 }
             )
 
-            const user = response.data
-
-            const userString = JSON.stringify(user)
+            const userString = JSON.stringify(userData)
             SecureStore.setItem("currentUser", userString)
-            axios.defaults.headers.common['Authorization'] = `Bearer ${user.api_key}`
+            axios.defaults.headers.common['Authorization'] = `Bearer ${userData.api_key}`
 
+            setUser(userData)
 
-            setUser(user)
-
-
-        } catch (error) {
-            console.error(JSON.stringify(error.response.data, null, 2));
+        } catch ({ response, request }) {
+            if (response) {
+                Alert.alert('Registration Failed', response.data.error)
+            } else if (request) {
+                Alert.alert('Connection Failed', 'please check your internet connection and try again.')
+            } else {
+                console.log('Error', error.message);
+            }
         }
-
     }
 
 
@@ -43,6 +48,9 @@ export const AuthContextProvider = ({ children }) => {
                 {
                     email,
                     password,
+                },
+                {
+                    timeout: 5000
                 }
             )
 
@@ -53,14 +61,32 @@ export const AuthContextProvider = ({ children }) => {
 
             setUser(user)
 
-        } catch (error) {
-            console.log(JSON.stringify(error.response.data, null, 2));
+        } catch ({ response, request }) {
+            if (response) {
+                Alert.alert('Login Failed', 'please check your username or password and try again.')
+            } else if (request) {
+                Alert.alert('Connection Failed', 'please check your internet connection and try again.')
+            } else {
+                console.log('Error', error.message);
+            }
         }
     }
 
     const handleLogout = () => {
-        SecureStore.deleteItemAsync("currentUser")
-        setUser(null)
+        Alert.alert("You are about to logout", "continue?", [
+            {
+                text: "Yes",
+                onPress: () => {
+                    SecureStore.deleteItemAsync("currentUser")
+                    setUser(null)
+                },
+            },
+            {
+                text: "No",
+                isPreferred : true,
+                style: 'cancel'
+            },
+        ])
     }
 
     useEffect(() => {
